fix(e2e): tree-expand + kebab-close fixes for mapping history (#313)
The PR-C1 spec inherited two bugs that PR-A's fixes already addressed
for mapping-flow.spec.ts but that did not propagate when the Phase 3
agents wrote new specs:

1. openRequirementDetail clicked the leaf row's text directly without
   first expanding its ancestor categories. The seeded catalogs (SOC 2 /
   ISO 27001 / HIPAA) have ONLY category rows at the top level — leaves
   are nested 1+ levels deep and hidden until each ancestor's chevron is
   clicked. Every UI test that walked through openRequirementDetail
   timed out.

2. FrameworkDetail's Edit/Delete kebab handlers in this branch did not
   close the menu (only the new History handler did). With a 3-item
   kebab the dropdown overlaps the inline confirm dialog that opens on
   Delete, and the still-open Delete <button role='menuitem'> intercepts
   pointer events on the Confirm button.

Fixes:
- Extract the recursive leaf-discovery + ancestor-expanding logic into
  frontend/e2e/_mapping-helpers.ts so future specs reuse it instead of
  copy-pasting.
- Have mapping-history.spec.ts import discoverMappingFixture +
  openRequirementDetail from the helper.
- Add setOpenMenuId(null) to the Edit and Delete handlers in
  FrameworkDetail.tsx alongside the existing History handler.
